Split tests: restore original test_simpledoc.py, add test_simpledoc_ci.py
- test_simpledoc.py: Original unit test (manual run with python -m)
- test_simpledoc_ci.py: Pytest CI tests with full round-trip + ResponseConverter
